backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
pydantic==2.5.3
opencv-python==4.9.0.80
mtcnn==0.1.1
numpy==1.26.3
pillow==10.2.0
bcrypt==4.1.2
python-dotenv==1.0.0
flask-limiter==3.5.0
pandas==2.1.4
pytest==7.4.3
gunicorn==21.2.0
python-dotenv>=1.0.0
